refactoring to make TaskRepository independent from DB implementation so I can change DB to other one in the feature

+++ OfcourseTodoApp/Data/Local/CoreData/CoreDataTaskDatabase.swift
@@ -8,30 +8,38 @@
 import Foundation
 import CoreData
 
-protocol TaskRepositoryProtocol {
-    @discardableResult func saveTask(title: String, comment: String?, isCompleted: Bool) -> Task?
-    func getTask(by id: NSManagedObjectID) -> Result<Task, Error>
-    func updateTask(with id: NSManagedObjectID, title: String, comment: String?, isCompleted: Bool?) -> Task?
-    func removeTask(by id: NSManagedObjectID) -> Result<Bool, Error>
+protocol TaskDatabaseProtocol {
+    @discardableResult 
+    func saveTask(title: String, comment: String?, isCompleted: Bool) -> Task?
+    func updateTask(with id: UUID, title: String, comment: String?, isCompleted: Bool?) -> Task?
+    func removeTask(by id: UUID) -> Result<Bool, Error>
+    
+    func fetchTask(by id: UUID) -> Result<Task, Error>
     func fetchAllTasks() -> [Task]
     func fetchIncompleteTasks() -> [Task]
+    
     func searchTasks(by title: String, includeOnlyIncomplete: Bool) -> [Task]
 }
 
-class CoreDataTaskRepository: TaskRepositoryProtocol {
+class CoreDataTaskDatabase: TaskDatabaseProtocol {
 
     private let persistentContainer: NSPersistentContainer
     
     init(persistentContainer: NSPersistentContainer) {
         self.persistentContainer = persistentContainer
     }
-    
-    func getTask(by id: NSManagedObjectID) -> Result<Task, Error> {
+
+    func fetchTask(by id: UUID) -> Result<Task, Error> {
         let context = persistentContainer.viewContext
+        let fetchRequest: NSFetchRequest<TaskEntity> = TaskEntity.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "uuid == %@", id as CVarArg)
 
         do {
-            let taskEntity = try context.existingObject(with: id) as! TaskEntity
-            return .success(TaskMapper.mapToObject(taskEntity: taskEntity))
+            if let taskEntity = try context.fetch(fetchRequest).first {
+                return .success(TaskMapper.mapToObject(taskEntity: taskEntity))
+            } else {
+                return .failure(NSError(domain: "Task not found", code: 404, userInfo: nil))
+            }
         } catch {
             print("Error fetching task: \(error.localizedDescription)")
             return .failure(error)
@@ -40,10 +48,10 @@ class CoreDataTaskRepository: TaskRepositoryProtocol {
 
     @discardableResult
     func saveTask(title: String, comment: String?, isCompleted: Bool = false) -> Task? {
-        
         let context = persistentContainer.viewContext
         let taskEntity = TaskEntity(context: context)
         
+        taskEntity.uuid = UUID()
         taskEntity.title = title
         taskEntity.comment = comment
         taskEntity.dateCreated = Date()
@@ -58,31 +66,41 @@ class CoreDataTaskRepository: TaskRepositoryProtocol {
         }
     }
 
-    func updateTask(with id: NSManagedObjectID, title: String, comment: String?, isCompleted: Bool?) -> Task? {
+    func updateTask(with id: UUID, title: String, comment: String?, isCompleted: Bool?) -> Task? {
         let context = persistentContainer.viewContext
+        let fetchRequest: NSFetchRequest<TaskEntity> = TaskEntity.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "uuid == %@", id as CVarArg)
 
         do {
-            let taskEntity = try context.existingObject(with: id) as! TaskEntity
-            taskEntity.title = title
-            taskEntity.comment = comment
-            taskEntity.isCompleted = isCompleted ?? taskEntity.isCompleted
-            
-            try saveContext()
-            return TaskMapper.mapToObject(taskEntity: taskEntity)
+            if let taskEntity = try context.fetch(fetchRequest).first {
+                taskEntity.title = title
+                taskEntity.comment = comment
+                taskEntity.isCompleted = isCompleted ?? taskEntity.isCompleted
+
+                try saveContext()
+                return TaskMapper.mapToObject(taskEntity: taskEntity)
+            } else {
+                return nil
+            }
         } catch {
             print("Failed to update task: \(error)")
             return nil
         }
     }
-    
-    func removeTask(by id: NSManagedObjectID) -> Result<Bool, Error> {
+
+    func removeTask(by id: UUID) -> Result<Bool, Error> {
         let context = persistentContainer.viewContext
+        let fetchRequest: NSFetchRequest<TaskEntity> = TaskEntity.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "uuid == %@", id as CVarArg)
 
         do {
-            let taskEntity = try context.existingObject(with: id) as! TaskEntity
-            context.delete(taskEntity)
-            try saveContext()
-            return .success(true)
+            if let taskEntity = try context.fetch(fetchRequest).first {
+                context.delete(taskEntity)
+                try saveContext()
+                return .success(true)
+            } else {
+                return .failure(NSError(domain: "Task not found", code: 404, userInfo: nil))
+            }
         } catch {
             print("Error fetching or deleting task: \(error.localizedDescription)")
             return .failure(error)
@@ -92,9 +110,10 @@ class CoreDataTaskRepository: TaskRepositoryProtocol {
     func fetchAllTasks() -> [Task] {
         let context = persistentContainer.viewContext
         let fetchRequest: NSFetchRequest<TaskEntity> = TaskEntity.fetchRequest()
+        
         do {
-            let taskEntity = try context.fetch(fetchRequest)
-            return taskEntity.map { TaskMapper.mapToObject(taskEntity: $0) }
+            let taskEntities = try context.fetch(fetchRequest)
+            return taskEntities.map { TaskMapper.mapToObject(taskEntity: $0) }
         } catch {
             print("Failed to fetch tasks: \(error)")
             return []
@@ -121,7 +140,6 @@ class CoreDataTaskRepository: TaskRepositoryProtocol {
         let fetchRequest: NSFetchRequest<TaskEntity> = TaskEntity.fetchRequest()
 
         var predicates: [NSPredicate] = []
-
         predicates.append(NSPredicate(format: "title CONTAINS[cd] %@", title))
 
         if includeOnlyIncomplete {
@@ -139,7 +157,6 @@ class CoreDataTaskRepository: TaskRepositoryProtocol {
         }
     }
 
-    
     private func saveContext() throws {
         let context = persistentContainer.viewContext
         if context.hasChanges {
@@ -147,4 +164,3 @@ class CoreDataTaskRepository: TaskRepositoryProtocol {
         }
     }
 }
-
